Is help a "Character"? I dunno. But it seems to belong here.
