+++ Characters/init.lua
@@ -1,2 +1,3 @@
+require "Characters/Help"
 require "Characters/Player1"
 require "Characters/Player2"
